requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
mcp[cli]>=1.3.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.120.0
pyngrok>=7.2.0
httpx>=0.27.0
python-multipart>=0.0.9
anyio>=4.4.0
aiofiles>=23.2.0
python-dotenv>=1.0.0
